Updating savings
I'm not sure if you had the version where i had changed the savings
decision. As I said, i'm not sure if it is right now, or is I should
have done something else when I was the consumption choice way.

I also commented out the line:
if (Age == Age0) return
!any(Alpha[][attend.pos~work.pos~savings.pos~GrowUp.pos]);

As it was giving me an error in the create spaces call.

I'll continue to trim the reachable state space, as I realize that there
are unreachable states i haven't trimmed.