RandomUpDown/UncertainLongevity
Here are the files related to that "issue" i posted on github.

I tried to write a simplified version of the model to figure out what is
going on in the main version. I basically set up a model of work, where
agents are supposed to randomly accumulate human capital during the age
phase of the model, then have constant human capital after. There is no
savings, you just consume your wages.

I've saved the output as "Ox Output - RandomUpDown". Basically if I run
the variable with constant human capital my results in the "age phase"
the model make sense. However, as soon as I add uncertainty about human
capital I begin to "lose" people in my distributions.

As I said in the github, it could be because I am not adjusting the
probabilities based on whether the person is at the max or min of the
state variable. I am going to try adjusting the probabilities based on
the value of the state variable to see if that is the problem.