Simple Program
I tried to simplify the program to see how PermanentChoiceWithReset is
working.

It's a simple program with both a PermanentChoice variable (GROWNUp) and
a PermanentChoiceWithReset variable (SchoolType).

Now, I get the opposite of the problem I was getting before, with
histograms adding up to more than one. Also, I need a certain line in
reachable in order for the transitions of the new state variable to run.
I'm trying to figure out why.