Data etc.
Here is a stata data file with the variables in the model, I am still
trying to figure out exactly how to deal with multiple school types
(they need to be constant in my model) and Credits (there are both BA
credits and associate credits).

For missing values of testscores used predicted probability from an
ordered probit. I averaged across value of parental wealth and income
(as does most of the literature).

I'll compute some moments using the egen and by commands today and
tomorrow and upload them. As well as clean up the data a bit more. I
will also use the rust example to get the data into niqlow.